requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas>=2.1.4
numpy>=1.24.3
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
schedule>=1.2.0
firecrawl-py>=0.0.16
python-dateutil>=2.8.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
redis>=5.0.0
aioredis>=2.0.1
celery>=5.3.0
celery[redis]>=5.3.0